+++ mex/common/models/access_platform.py
@@ -65,7 +65,7 @@ class _SparseValues(_Stem):
 
 
 class _VariadicValues(_Stem):
-    endpointDescription: list[Link]
+    endpointDescription: list[Link] = []
     endpointType: list[APIType] = []
     endpointURL: list[Link] = []
     technicalAccessibility: list[TechnicalAccessibility] = []
@@ -141,9 +141,9 @@ class PreventiveAccessPlatform(_Stem, PreventiveRule):
 
 
 class _BaseRuleSet(_Stem, RuleSet):
-    additive: AdditiveAccessPlatform
-    subtractive: SubtractiveAccessPlatform
-    preventive: PreventiveAccessPlatform
+    additive: AdditiveAccessPlatform = AdditiveAccessPlatform()
+    subtractive: SubtractiveAccessPlatform = SubtractiveAccessPlatform()
+    preventive: PreventiveAccessPlatform = PreventiveAccessPlatform()
 
 
 class AccessPlatformRuleSetRequest(_BaseRuleSet):