+++ mex/common/types/__init__.py
@@ -75,6 +75,23 @@ from mex.common.types.vocabulary import (
 )
 
 __all__ = (
+    "CET",
+    "EMAIL_PATTERN",
+    "EXTRACTED_IDENTIFIER_CLASSES",
+    "EXTRACTED_IDENTIFIER_CLASSES_BY_NAME",
+    "IDENTIFIER_PATTERN",
+    "MERGED_IDENTIFIER_CLASSES",
+    "MERGED_IDENTIFIER_CLASSES_BY_NAME",
+    "NESTED_MODEL_CLASSES",
+    "NESTED_MODEL_CLASSES_BY_NAME",
+    "TEMPORAL_ENTITY_CLASSES_BY_PRECISION",
+    "TEMPORAL_ENTITY_FORMATS_BY_PRECISION",
+    "URL_PATTERN",
+    "UTC",
+    "VOCABULARY_ENUMS",
+    "VOCABULARY_ENUMS_BY_NAME",
+    "VOCABULARY_PATTERN",
+    "APIType",
     "AccessRestriction",
     "ActivityType",
     "AnonymizationPseudonymization",
@@ -82,18 +99,14 @@ __all__ = (
     "AnyMergedIdentifier",
     "AnyNestedModel",
     "AnyPrimitiveType",
-    "APIType",
+    "AnyVocabularyEnum",
     "AssetsPath",
     "BibliographicResourceType",
-    "CET",
     "ConsentStatus",
     "ConsentType",
     "DataProcessingState",
     "DataType",
-    "EMAIL_PATTERN",
     "Email",
-    "EXTRACTED_IDENTIFIER_CLASSES_BY_NAME",
-    "EXTRACTED_IDENTIFIER_CLASSES",
     "ExtractedAccessPlatformIdentifier",
     "ExtractedActivityIdentifier",
     "ExtractedBibliographicResourceIdentifier",
@@ -101,15 +114,14 @@ __all__ = (
     "ExtractedContactPointIdentifier",
     "ExtractedDistributionIdentifier",
     "ExtractedIdentifier",
-    "ExtractedOrganizationalUnitIdentifier",
     "ExtractedOrganizationIdentifier",
+    "ExtractedOrganizationalUnitIdentifier",
     "ExtractedPersonIdentifier",
     "ExtractedPrimarySourceIdentifier",
     "ExtractedResourceIdentifier",
     "ExtractedVariableGroupIdentifier",
     "ExtractedVariableIdentifier",
     "Frequency",
-    "IDENTIFIER_PATTERN",
     "Identifier",
     "IdentityProvider",
     "Language",
@@ -117,8 +129,7 @@ __all__ = (
     "Link",
     "LinkLanguage",
     "LiteralStringType",
-    "MERGED_IDENTIFIER_CLASSES_BY_NAME",
-    "MERGED_IDENTIFIER_CLASSES",
+    "MIMEType",
     "MergedAccessPlatformIdentifier",
     "MergedActivityIdentifier",
     "MergedBibliographicResourceIdentifier",
@@ -126,33 +137,24 @@ __all__ = (
     "MergedContactPointIdentifier",
     "MergedDistributionIdentifier",
     "MergedIdentifier",
-    "MergedOrganizationalUnitIdentifier",
     "MergedOrganizationIdentifier",
+    "MergedOrganizationalUnitIdentifier",
     "MergedPersonIdentifier",
     "MergedPrimarySourceIdentifier",
     "MergedResourceIdentifier",
     "MergedVariableGroupIdentifier",
     "MergedVariableIdentifier",
-    "MIMEType",
-    "NESTED_MODEL_CLASSES_BY_NAME",
-    "NESTED_MODEL_CLASSES",
     "PathWrapper",
     "PersonalData",
     "ResourceCreationMethod",
     "ResourceTypeGeneral",
     "Sink",
-    "split_to_caps",
     "TechnicalAccessibility",
-    "TEMPORAL_ENTITY_CLASSES_BY_PRECISION",
-    "TEMPORAL_ENTITY_FORMATS_BY_PRECISION",
     "TemporalEntity",
     "TemporalEntityPrecision",
     "Text",
     "TextLanguage",
     "Theme",
-    "URL_PATTERN",
-    "UTC",
-    "VOCABULARY_PATTERN",
     "VocabularyEnum",
     "VocabularyLoader",
     "WorkPath",
@@ -160,7 +162,42 @@ __all__ = (
     "YearMonth",
     "YearMonthDay",
     "YearMonthDayTime",
+    "split_to_caps",
 )
+
+AnyVocabularyEnum = (
+    AccessRestriction
+    | ActivityType
+    | AnonymizationPseudonymization
+    | APIType
+    | BibliographicResourceType
+    | ConsentStatus
+    | ConsentType
+    | DataProcessingState
+    | Frequency
+    | Language
+    | License
+    | MIMEType
+    | PersonalData
+    | ResourceCreationMethod
+    | ResourceTypeGeneral
+    | TechnicalAccessibility
+    | Theme
+)
+VOCABULARY_ENUMS: Final[list[type[AnyVocabularyEnum]]] = list(
+    get_args(AnyVocabularyEnum)
+)
+VOCABULARY_ENUMS_BY_NAME: Final[dict[str, type[AnyVocabularyEnum]]] = {
+    cls.__name__: cls for cls in VOCABULARY_ENUMS
+}
+
+AnyTemporalEntity = Year | YearMonth | YearMonthDay | YearMonthDayTime
+TEMPORAL_ENTITIES: Final[list[type[AnyTemporalEntity]]] = list(
+    get_args(AnyTemporalEntity)
+)
+TEMPORAL_ENTITIES_BY_NAME: Final[dict[str, type[AnyTemporalEntity]]] = {
+    cls.__name__: cls for cls in TEMPORAL_ENTITIES
+}
 
 AnyNestedModel = Link | Text
 NESTED_MODEL_CLASSES: Final[list[type[AnyNestedModel]]] = list(get_args(AnyNestedModel))